requirements.txt
fastapi
uvicorn
mem0ai
psycopg2-binary
sqlalchemy
ollama
pydantic
alembic
python-dotenv
pytest
httpx
qdrant-client